+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `white-space: nowrap` added to `<code>` tags on posts
 - Proper title to Blog page
 - Proper title to Posts pages
+- Like Button on post single pages
 
 ### CHANGED